AbstractWakeupMote: implement WatchpointMote
Most of the WatchpointMote functionality is
completely generic, so implement that in
AbstractWakeupMote so all motes can re-use
the functionality.

Motes that support watchpoints currently
need to override addBreakpoint, setConfigXML(),
and getExecutableAddressOf(), and add an implements
for WatchpointMote. See MspMote for an example.
